+++ localization/Localization/ArticleManager.cs
@@ -1,0 +1,6 @@
+using Microsoft.Extensions.Localization;
+
+public class ArticleManager(IStringLocalizer<ArticleManager> _localizer)
+{
+    public string GetArticleName() => _localizer["articleName"];
+}
