+++ localization/Localization/Localization.csproj
@@ -1,6 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Localization" />
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" />
   </ItemGroup>
