+++ localization/Localization/Program.cs
@@ -1,0 +1,37 @@
+using System.Globalization;
+using Microsoft.AspNetCore.Localization;
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddOpenApi();
+builder.Services.AddLocalization(options => options.ResourcesPath = "Resources");
+builder.Services.Configure<RequestLocalizationOptions>(
+    options =>
+    {
+        var supportedCultures = new List<CultureInfo>
+        {
+            new CultureInfo("en-US"),
+            new CultureInfo("tr-TR")
+        };
+
+        options.DefaultRequestCulture = new RequestCulture(culture: "en-US", uiCulture: "en-US");
+        options.SupportedCultures = supportedCultures;
+    });
+
+builder.Services.AddSingleton<ArticleManager>();
+
+var app = builder.Build();
+
+app.UseRequestLocalization();
+
+if (app.Environment.IsDevelopment())
+{
+    app.MapOpenApi();
+}
+
+app.UseHttpsRedirection();
+
+var articleManager = app.Services.GetService<ArticleManager>()!;
+app.MapGet("/", () => articleManager.GetArticleName());
+
+app.Run();
