+++ localization/Localization/Properties/launchSettings.json
