+++ localization/Localization/localization.http
@@ -1,0 +1,3 @@
+@Localization_HostAddress = http://localhost:5292
+
+###
