--- localization/Program.cs
@@ -1,16 +1,0 @@
-var builder = WebApplication.CreateBuilder(args);
-
-builder.Services.AddOpenApi();
-
-var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.MapOpenApi();
-}
-
-app.UseHttpsRedirection();
-
-app.MapGet("/", () => "Hello World!");
-
-app.Run();
