--- localization/localization.http
@@ -1,6 +1,0 @@
-@localization_HostAddress = http://localhost:5292
-
-GET {{localization_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
